@@ -1,7 +1,7 @@
 {
-  "name": "gatsby-theme-lumen",
   "version": "1.0.0",
-  "description": "Awesome Theme for Gatsby",
+  "name": "gatsby-starter-lumen",
+  "description": "Simple starter for Gatsby",
   "main": "n/a",
   "scripts": {
     "lint": "./node_modules/.bin/eslint --ext .js,.jsx --ignore-pattern public .",
@@ -9,41 +9,40 @@
   },
   "keywords": [
     "gatsby",
-    "theme",
-    "lumen"
+    "lumen",
+    "starter"
   ],
   "author": "Alexander Shelepenok <a.shelepenok@gmail.com>",
   "license": "MIT",
   "dependencies": {
-    "react": "^0.14.7",
+    "babel-plugin-add-module-exports": "^0.2.0",
+    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-react": "^6.5.0",
+    "babel-preset-stage-0": "^6.5.0",
+    "babel-preset-stage-1": "^6.5.0",
+    "classnames": "^2.1.2",
+    "front-matter": "^2.0.8",
+    "gatsby": "^0.11.2",
+    "history": "^2.1.1",
+    "lodash": "^4.12.0",
+    "moment": "^2.12.0",
+    "markdown-it": "^6.0.2",
     "react-document-title": "^2.0.1",
-    "react-dom": "^0.14.7",
-    "react-router": "^2.0.0",
-    "lodash": "^4.5.0",
+    "react-google-analytics": "^0.2.0",
+    "react-router": "^2.4.0",
     "safe-access": "^0.1.0",
-    "underscore.string": "^3.2.3",
-    "moment": "^2.11.2"
+    "underscore.string": "^3.2.3"
   },
   "devDependencies": {
-    "webpack": "^1.12.2",
-    "webpack-dev-server": "^1.12.0",
-    "webpack-hot-middleware": "^2.2.0",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.8.4",
-    "url-loader": "^0.5.7",
-    "sugarss": "^0.1.0",
-    "precss": "^1.4.0",
-    "postcss-scss": "^0.1.7",
-    "postcss-loader": "^0.8.0",
-    "postcss-modules": "^0.1.3",
-    "lost": "^6.7.0",
+    "autoprefixer": "^6.3.6",
+    "file-loader": "^0.8.5",
     "rucksack-css": "^0.8.5",
-    "babel-plugin-react-transform": "^1.1.1",
-    "eslint": "^2.3.0",
-    "eslint-config-airbnb": "^6.1.0",
-    "eslint-plugin-react": "^4.2.0",
-    "react-transform-catch-errors": "^1.0.0",
-    "react-transform-hmr": "^1.0.0",
-    "redbox-react": "^1.0.1"
+    "lost": "^6.7.2",
+    "babel-preset-react-hmre": "^1.1.1",
+    "eslint": "^2.9.0",
+    "eslint-config-airbnb": "^9.0.1",
+    "eslint-plugin-import": "^1.8.0",
+    "eslint-plugin-jsx-a11y": "^1.2.0",
+    "eslint-plugin-react": "^5.1.1"
   }
 }
