@@ -1,27 +1,30 @@
 import React from 'react'
 import { RouteHandler, Link } from 'react-router'
-import { link } from 'gatsby-helpers'
+import { prefixLink } from 'gatsby-helpers'
 import { config } from 'config'
 import BlogNav from '../BlogNav'
 import BlogSocial from '../BlogSocial'
-import './style.sss'
+import './style.css'
 
 class SidebarLeft extends React.Component {
     render() {
         const {location, children} = this.props
-        let header
+        const isHome = location.pathname === prefixLink('/')
 
-        header = (
-            <header>
-              <Link style={ {    textDecoration: 'none',    borderBottom: 'none',    outline: 'none'} } to={ link('/') }>
-              <img src='./images/photo9.jpg' width='75' height='75' />
-              </Link>
-              <h1><Link style={ {    textDecoration: 'none',    borderBottom: 'none',    color: 'inherit'} } to={ link('/') } > { config.siteTitle } </Link></h1>
-              <p>
-                { config.siteDescr }
-              </p>
-            </header>
-        );
+        let header = (
+        <header>
+          <Link style={ {    textDecoration: 'none',    borderBottom: 'none',    outline: 'none'} } to={ prefixLink('/') }>
+          <img src='./photo.jpg' width='75' height='75' />
+          </Link>
+          { isHome ? (
+            <h1><Link style={ {    textDecoration: 'none',    borderBottom: 'none',    color: 'inherit'} } to={ prefixLink('/') }> { config.siteAuthor } </Link></h1>
+            ) :
+            <h2><Link style={ {    textDecoration: 'none',    borderBottom: 'none',    color: 'inherit'} } to={ prefixLink('/') }> { config.siteAuthor } </Link></h2> }
+          <p>
+            { config.siteDescr }
+          </p>
+        </header>
+        )
 
         return (
             <div className='sidebar'>
